pyproject.toml
[project]
name = "docmost2outline"
version = "0.1.0"
description = "A tool to migrate documents from DocMost to Outline."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "python-magic>=0.4.27",
    "rich>=14.2.0",
    "tenacity>=9.1.2",
]
